nov 3, 22 daniel is gae and stupid
