+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmodes/ArmControlOpMode.java
@@ -4,6 +4,8 @@ import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 import com.qualcomm.robotcore.hardware.DcMotor;
 
+import org.firstinspires.ftc.teamcode.wrappers.Gear;
+import org.firstinspires.ftc.teamcode.wrappers.MotorRatio;
 import org.firstinspires.ftc.teamcode.wrappers.MotorWrapper;
 
 @TeleOp(name="ArmTestOpMode")
@@ -17,7 +19,7 @@ public class ArmControlOpMode extends LinearOpMode {
     public void runOpMode() throws InterruptedException {
         // setup
         arm = new MotorWrapper(hardwareMap.get(DcMotor.class, "m1"), 2.0, 0,
-                new MotorWrapper.MotorRatio(new MotorWrapper.Gear[]{new MotorWrapper.Gear(MotorWrapper.TICKS_TORQNADO)}));
+                new MotorRatio(new Gear[]{new Gear(MotorWrapper.TICKS_TORQNADO)}));
         // prerun
         waitForStart();
         arm.setTargetPower(0.4);
