+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmodes/OverallSample.java
@@ -5,6 +5,8 @@ import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.hardware.DcMotorSimple;
 
+import org.firstinspires.ftc.teamcode.wrappers.Gear;
+import org.firstinspires.ftc.teamcode.wrappers.MotorRatio;
 import org.firstinspires.ftc.teamcode.wrappers.MotorWrapper;
 
 @TeleOp(name="SampleTest")
@@ -32,7 +34,7 @@ public class OverallSample extends LinearOpMode {
 
         // setup
         arm = new MotorWrapper(hardwareMap.get(DcMotor.class, "m1"), 2.0, 0,
-                new MotorWrapper.MotorRatio(new MotorWrapper.Gear[]{new MotorWrapper.Gear(MotorWrapper.TICKS_TORQNADO)}));
+                new MotorRatio(new Gear[]{new Gear(MotorWrapper.TICKS_TORQNADO)}));
         // prerun
         waitForStart();
         arm.setTargetPower(0.4);
