+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/wrappers/Gear.java
@@ -1,0 +1,12 @@
+package org.firstinspires.ftc.teamcode.wrappers;
+
+public class Gear{
+    public int teeth;
+    public Gear(int teeth){
+        this.teeth = teeth;
+    }
+
+    public double convertRatio(Gear next, double spinTicks){
+        return (double)teeth / (double)next.teeth * (double)spinTicks;
+    }
+}
